bug fixed

+++ pyrlprob/base_funs.py
@@ -212,7 +212,7 @@ def evaluation(trainer: Union[str, Callable, Type],
 
     # Evaluation and callbacks config
     config["evaluation_interval"] = 1
-    config["evaluation_num_episodes"] = evaluation_num_episodes/config["num_envs_per_worker"]
+    config["evaluation_num_episodes"] = int(evaluation_num_episodes/config["num_envs_per_worker"])
     config["evaluation_config"] = evaluation_config
     if custom_eval_function is not None:
         if callable(custom_eval_function):
